split logic in different files
